mcp_server/requirements.txt
flask==2.3.3
aioflask==0.4.0
pyjwt==2.8.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
numpy==1.25.2
aiohttp==3.9.1
python-dateutil==2.8.2
plotly==5.17.0 